// src/global.d.ts
import { interpolateAs } from "next/dist/shared/lib/router/router";

export interface User {
  country: string;
  display_name: string;
  email: string;
  explicit_content: ExplicitContent;
  external_urls: ExternalUrls;
  followers: Followers;
  href: string;
  id: string;
  images: Image[];
  product: string;
  type: string;
  uri: string;
}

export interface ExplicitContent {
  filter_enabled: boolean;
  filter_locked: boolean;
}

export interface ExternalUrls {
  spotify: string;
}

export interface Followers {
  href: null;
  total: number;
}

export interface Image {
  height: null;
  url: string;
  width: null;
}

// Generated by https://quicktype.io

export interface Sticker {
  id: number;
  attributes: StickerAttributes;
}

export interface StickerAttributes {
  createdAt: string;
  updatedAt: string;
  image: Image;
  category: Category;
}

export interface Category {
  data: CategoryData;
}

export interface CategoryData {
  id: number;
  attributes: PurpleAttributes;
}

export interface PurpleAttributes {
  name: string;
  createdAt: string;
  updatedAt: string;
}

export interface Image {
  data: ImageData;
}

export interface ImageData {
  id: number;
  attributes: FluffyAttributes;
}

export interface FluffyAttributes {
  name: string;
  alternativeText: null;
  caption: null;
  width: number;
  height: number;
  formats: Formats;
  hash: string;
  ext: string;
  mime: string;
  size: number;
  url: string;
  previewUrl: null;
  provider: string;
  provider_metadata: null;
  createdAt: string;
  updatedAt: string;
}

export interface Formats {
  thumbnail: Thumbnail;
}

export interface Thumbnail {
  name: string;
  hash: string;
  ext: string;
  mime: string;
  path: null;
  width: number;
  height: number;
  size: number;
  url: string;
}

// Generated by https://quicktype.io

export interface UserSticker {
  id: number;
  attributes: UserStickerAttributes;
}

export interface UserStickerAttributes {
  email: string;
  createdAt: string;
  updatedAt: string;
  sticker: Sticker;
}

export interface Sticker {
  data: Data;
}

export interface Data {
  id: number;
  attributes: DataAttributes;
}

export interface DataAttributes {
  createdAt: string;
  updatedAt: string;
}